Move character model to character app
