+++ character/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import Character
+
+
+class CharacterSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Character
+        fields = ['id', 'name', 'xp', 'xp_next_level', 'xp_modifier', 'level', 'total_quests']
