+++ gameplay/serializers.py
@@ -1,11 +1,6 @@
 from rest_framework import serializers
-from .models import Quest, Activity, Character, QuestResults
+from .models import Quest, Activity, QuestResults
 
-
-class CharacterSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Character
-        fields = ['id', 'name', 'xp', 'xp_next_level', 'xp_modifier', 'level', 'total_quests']
 
 class QuestResultSerializer(serializers.ModelSerializer):
     class Meta:
