+++ progress_rpg/settings.py
@@ -99,6 +99,8 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'django_celery_beat',
     'rest_framework',
+
+    'character',
     'users',
     'gameplay',
     'gameworld',
