+++ users/signals.py
@@ -1,7 +1,8 @@
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from .models import Profile
-from gameplay.models import Character, DailyStats
+from gameplay.models import DailyStats
+from character.models import Character
 from django.contrib.auth import get_user_model
 from gameplay.utils import assign_character_to_profile
